requirements.txt
-c constraints.txt

Django
djangorestframework
psycopg2
requests
